update submodule imgui and fix warnings.

+++ Ether/src/Ether/Renderer/Renderer2D.cpp
@@ -140,7 +140,7 @@ namespace Ether
 	{
 		ETHER_PROFILE_FUNCTION();
 
-		uint32_t data_size = (uint8_t*)s_Data.QuadVertexBufferPtr - (uint8_t*)s_Data.QuadVertexBufferBase;
+		uint32_t data_size = (uint32_t)((uint8_t*)s_Data.QuadVertexBufferPtr - (uint8_t*)s_Data.QuadVertexBufferBase);
 		s_Data.QuadVertexBuffer->SetData(s_Data.QuadVertexBufferBase, data_size);
 
 		Flush();
